chapter-08/requirements_weavite.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
asyncer==0.0.8
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
cryptography==46.0.3
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dspy==3.0.4
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
gepa==0.0.17
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json_repair==0.54.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.80.9
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.5
openai==2.11.0
optuna==4.6.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
validators==0.35.0
weaviate-client==4.18.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
